Issues presented by ebubar resolved
ebubar:

"The Arduino.ino file has an extra semi-colon after the include servo. "

"Just a minor spacing issue in arduino_port.py on line 25. My compiling
complained about the spaces before the "if" statement. Commenting the
line above with a # and getting rid of space before "if" seems to fix
the problem for me."